// apps/kampus/app/pano/__generated__/SinglePostFeed_comments.graphql.ts
export type SinglePostFeed_comments$data = {
  readonly comments: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly content: string;
        readonly createdAt: any;
        readonly id: string;
        readonly owner: {
          readonly displayName: string | null;
        } | null;
        readonly " $fragmentSpreads": FragmentRefs<"CommentItem_comment">;
      } | null;
    } | null> | null;
  } | null;
  readonly id: string;
  readonly " $fragmentType": "SinglePostFeed_comments";
};
export type SinglePostFeed_comments$key = {
  readonly " $data"?: SinglePostFeed_comments$data;
  readonly " $fragmentSpreads": FragmentRefs<"SinglePostFeed_comments">;
};